+++ src/app/features/products/components/product-view/product-view.component.ts
@@ -1,12 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute, ParamMap } from '@angular/router';
 import { Store } from '@ngrx/store';
-import { Observable } from 'rxjs';
-import { switchMap } from 'rxjs/operators';
 import { ShpState } from 'src/app/core/@ngrx';
-import { getProduct, getProductSuccess, selectEditedProductByUrl } from 'src/app/core/@ngrx/products';
+import { selectEditedProductByUrl } from 'src/app/core/@ngrx/products';
 import { ProductModel } from '../../models/product';
-import { ProductsPromiseService } from '../../services/products-promise.service';
 
 @Component({
   selector: 'shp-product-view',
@@ -18,7 +14,6 @@ export class ProductViewComponent implements OnInit {
 
   constructor(
     private store: Store<ShpState>,
-    private route: ActivatedRoute,
   ) { }
 
   ngOnInit(): void {